Use Long for numberOfPropagations and similar stats

+++ jni/src/main/cpp/JGlucose.cpp
@@ -66,17 +66,17 @@ JNI_METHOD(jint, glucose_1nlearnts)
     return decode(handle)->nLearnts();
 }
 
-JNI_METHOD(jint, glucose_1decisions)
+JNI_METHOD(jlong, glucose_1decisions)
   (JNIEnv*, jobject, jlong handle) {
     return decode(handle)->decisions;
 }
 
-JNI_METHOD(jint, glucose_1propagations)
+JNI_METHOD(jlong, glucose_1propagations)
   (JNIEnv*, jobject, jlong handle) {
     return decode(handle)->propagations;
 }
 
-JNI_METHOD(jint, glucose_1conflicts)
+JNI_METHOD(jlong, glucose_1conflicts)
   (JNIEnv*, jobject, jlong handle) {
     return decode(handle)->conflicts;
 }
